orgazing imports and adding flex to expanded
