+++ lib/map/view/pages/map_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:free_fire_location/consts/colors.dart';
 import 'package:free_fire_location/map/view/cubit/fire/fire_cubit.dart';
-import 'package:free_fire_location/map/view/pages/splash.dart';
+import 'package:free_fire_location/map/view/pages/splash_page.dart';
 import 'package:free_fire_location/map/view/widgets/map_widget.dart';
 
 class MapPage extends StatelessWidget {
