+++ lib/map/view/widgets/map_widget.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:free_fire_location/map/view/cubit/fire/fire_cubit.dart';
-import 'package:free_fire_location/map/view/pages/splash.dart';
+import 'package:free_fire_location/map/view/pages/splash_page.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 class MapWidget extends StatefulWidget {
