+++ lib/map/view/widgets/menu_widget.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class MenuWidget extends StatelessWidget {
+  const MenuWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
