+++ lib/map/view/widgets/splash_widget.dart
