+++ lib/my_app.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:free_fire_location/map/view/cubit/fire/fire_cubit.dart';
-import 'package:free_fire_location/map/view/pages/map.dart';
-import 'package:free_fire_location/map/view/pages/splash.dart';
+import 'package:free_fire_location/map/view/pages/map_page.dart';
 
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
@@ -9,7 +9,6 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final fireCubit = FireCubit();
-
 
     return MultiBlocProvider(
       providers: [
